+++ Reach/test/utopia/reach/test/TestCursors.scala
@@ -18,14 +18,14 @@ object TestCursors
 {
 	private val cursorsDirectory: Path = "Reach/test-images"
 	lazy val cursors: Option[CursorSet] = Image.readFrom(cursorsDirectory/"cursor-arrow.png").toOption.map { arrowImage =>
-		val arrowCursor = Cursor(SingleColorIcon(arrowImage.withSourceResolutionOrigin(Point(7, 4))))
+		val arrowCursor = Cursor(SingleColorIcon(arrowImage.withSourceResolutionOrigin(Point(7, 4))), drawEdges = true)
 		val handImage = Image.readFrom(cursorsDirectory/"cursor-hand.png").toOption.map { i =>
 			SingleColorIcon(i.withSourceResolutionOrigin(Point(9, 1))) }
 		val textImage = Image.readFrom(cursorsDirectory/"cursor-text.png").toOption.map { i =>
 			SingleColorIcon(i.withCenterOrigin) }
 		
 		CursorSet(Vector(Interactive -> handImage, Text -> textImage)
-			.flatMap { case (cursorType, cursor) => cursor.map { cursorType -> Cursor(_) } }
+			.flatMap { case (cursorType, cursor) => cursor.map { cursorType -> Cursor(_, drawEdges = true) } }
 			.toMap[CursorType, Cursor] + (Default -> arrowCursor), arrowCursor)
 	}
 }